Use consistent interface name templates (fixes #723)
Also:
* Move string functions from 'common' to 'utils.strings'
* Move logging/error functions from 'common' to 'utils.log'
* Remove 'set_verbose' function that was used exactly once
* Fix way too many test cases ;)
