+++ netsim/devices/asav.yml
@@ -1,7 +1,7 @@
 #
 # Cisco ASAv
 #
-interface_name: GigabitEthernet0/%d
+interface_name: GigabitEthernet0/{ifindex}
 ifindex_offset: 0
 mgmt_if: Management0/0
 group_vars:
