+++ netsim/devices/csr.yml
@@ -1,7 +1,7 @@
 #
 # Cisco CSR 1000v
 #
-interface_name: GigabitEthernet%d
+interface_name: GigabitEthernet{ifindex}
 loopback_interface_name: Loopback{ifindex}
 ifindex_offset: 2
 virtualbox:
