+++ netsim/devices/cumulus.yml
@@ -1,7 +1,7 @@
 #
 # Cumulus VX 4.x or 5.x configured without NVUE
 #
-interface_name: swp%d
+interface_name: swp{ifindex}
 loopback_interface_name: "lo{ifindex}"
 mgmt_if: eth0
 libvirt:
