+++ netsim/devices/cumulus_nvue.yml
@@ -1,7 +1,7 @@
 #
 # Cumulus VX 5.x configured with NVUE
 #
-interface_name: swp%d
+interface_name: swp{ifindex}
 mgmt_if: eth0
 libvirt:
   image: CumulusCommunity/cumulus-vx:5.2.0
