+++ netsim/devices/dellos10.yml
@@ -1,7 +1,7 @@
 #
 # Dell OS10
 #
-interface_name: ethernet1/1/%d
+interface_name: ethernet1/1/{ifindex}
 mgmt_if: mgmt1/1/1
 loopback_interface_name: loopback{ifindex}
 features:
@@ -29,7 +29,7 @@ clab:
   node:
     kind: vr-ftosv
   interface:
-    name: eth%d
+    name: eth{ifindex}
 libvirt:
   image: dell/os10
   create:
