+++ netsim/devices/fortios.yml
@@ -1,7 +1,7 @@
 #
 # Fortinet FortiOS
 #
-interface_name: port%d
+interface_name: port{ifindex}
 mgmt_if: port1
 ifindex_offset: 2
 libvirt:
