+++ netsim/devices/frr.yml
@@ -1,7 +1,7 @@
 #
 # FRR container
 #
-interface_name: eth%d
+interface_name: eth{ifindex}
 mgmt_if: eth0
 loopback_interface_name: "lo{ifindex}"
 group_vars:
