+++ netsim/devices/iosv.yml
@@ -1,7 +1,7 @@
 #
 # Cisco IOSv
 #
-interface_name: GigabitEthernet0/%d
+interface_name: GigabitEthernet0/{ifindex}
 loopback_interface_name: Loopback{ifindex}
 group_vars:
   ansible_user: vagrant
