+++ netsim/devices/iosxr.yml
@@ -2,7 +2,7 @@
 # Cisco IOS XRv
 #
 mgmt_if: MgmtEth0/RP0/CPU0/0
-interface_name: GigabitEthernet0/0/0/%d
+interface_name: GigabitEthernet0/0/0/{ifindex}
 ifindex_offset: 0
 group_vars:
   ansible_user: vagrant
