+++ netsim/devices/linux.yml
@@ -1,7 +1,7 @@
 #
 # Generic Linux host
 #
-interface_name: eth%d
+interface_name: eth{ifindex}
 mgmt_if: eth0
 role: host
 libvirt:
