+++ netsim/devices/none.yml
@@ -4,7 +4,7 @@
 # Most features are enabled on the dummy device -- this file is a pretty
 # good template if you want to figure out what device features there are.
 #
-interface_name: eth%d
+interface_name: eth{ifindex}
 loopback_interface_name: Loopback{ifindex}
 virtualbox:
   image: none
