+++ netsim/devices/nxos.yml
@@ -1,6 +1,6 @@
 # Cisco Nexus 9300v
 #
-interface_name: Ethernet1/%d
+interface_name: Ethernet1/{ifindex}
 mgmt_if: mgmt0
 loopback_interface_name: loopback{ifindex}
 virtualbox:
