+++ netsim/devices/routeros.yml
@@ -1,6 +1,6 @@
 # Mikrotik RouterOS
 #
-interface_name: ether%d
+interface_name: ether{ifindex}
 mgmt_if: ether1
 ifindex_offset: 2
 libvirt:
