+++ netsim/devices/routeros7.yml
@@ -1,6 +1,6 @@
 # Mikrotik RouterOS version 7
 #
-interface_name: ether%d
+interface_name: ether{ifindex}
 mgmt_if: ether1
 loopback_interface_name: loopback{ifindex}
 ifindex_offset: 2
