+++ netsim/devices/srlinux.yml
@@ -1,7 +1,7 @@
 # Nokia SR Linux
 #
 mgmt_if: mgmt0
-interface_name: ethernet-1/%d
+interface_name: ethernet-1/{ifindex}
 loopback_interface_name: "lo0.{ifindex}"
 group_vars:
   ansible_user: admin
@@ -22,7 +22,7 @@ clab:
     kind: srl
     type: ixrd2
   interface:
-    name: e1-%d
+    name: e1-{ifindex}
   group_vars:
     srlinux_grpc_port: 57400
 features:
