+++ netsim/devices/sros.yml
@@ -1,7 +1,7 @@
 # Nokia SR OS
 #
 mgmt_if: A/1
-interface_name: 1/1/c%d
+interface_name: 1/1/c{ifindex}
 loopback_interface_name: "loopback.{ifindex}"
 group_vars:
   ansible_user: admin
@@ -64,7 +64,7 @@ clab:
     type: sr-1 # By default emulate SR-1
     license: /Projects/SR_OS_VSR-SIM_license.txt
   interface:
-    name: eth%d
+    name: eth{ifindex}
   group_vars:
     sros_grpc_port: 57400
 external:
