+++ netsim/devices/vsrx.yml
@@ -1,6 +1,6 @@
 # Juniper vSRX 3.0
 #
-interface_name: ge-0/0/%d
+interface_name: ge-0/0/{ifindex}
 loopback_interface_name: "lo0.{ifindex}"
 ifindex_offset: 0
 mgmt_if: fxp0
