+++ netsim/devices/vyos.yml
@@ -1,6 +1,6 @@
 # Vyatta VyOS VM/container
 #
-interface_name: eth%d
+interface_name: eth{ifindex}
 loopback_interface_name: "dum{ifindex}"
 mgmt_if: eth0
 libvirt:
