+++ netsim/utils/__init__.py
@@ -1,0 +1,4 @@
+# Import the whole augmentation tree
+
+from . import log
+from . import strings
