+++ tests/errors/ospf-unnumbered.yml
@@ -7,7 +7,7 @@ defaults:
     supported_on: [ iosv, fake_1, fake_2 ]
   devices:
     fake_1:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
       image:
         libvirt: cisco/iosv
         virtualbox: cisco/iosv
