+++ tests/errors/unnumbered-capability.yml
@@ -5,12 +5,12 @@
 defaults:
   devices:
     fake_1:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
       image:
         libvirt: cisco/iosv
         virtualbox: cisco/iosv
     fake_2:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
       image:
         libvirt: cisco/iosv
         virtualbox: cisco/iosv
