+++ tests/minimal_errors/addrs-mgmt-no-start.yml
@@ -11,7 +11,7 @@ defaults:
       config: Vagrantfile
   devices:
     iosv:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
       image:
         libvirt: x
