+++ tests/minimal_errors/bgp-no-defaults.yml
@@ -10,7 +10,7 @@ defaults:
       config: Vagrantfile
   devices:
     iosv:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
       image:
         libvirt: cisco/iosv
   bgp:
