+++ tests/minimal_errors/device-image-datatype.yml
@@ -12,7 +12,7 @@ defaults:
       config: Vagrantfile
   devices:
     iosv:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
       image: Wrong data type
 
 nodes:
