+++ tests/minimal_errors/device-image-missing.yml
@@ -12,7 +12,7 @@ defaults:
       config: Vagrantfile
   devices:
     iosv:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
 
 nodes:
   r1:
