+++ tests/minimal_errors/device-no-provider-image.yml
@@ -12,7 +12,7 @@ defaults:
       config: Vagrantfile
   devices:
     iosv:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
       image:
         virtualbox: x
