+++ tests/minimal_errors/no-mgmt-addr.yml
@@ -7,7 +7,7 @@ defaults:
       config: Vagrantfile
   devices:
     iosv:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
       image:
         libvirt: cisco/iosv
   bgp:
