+++ tests/topology/input/addrs-no-pool.yml
@@ -27,7 +27,7 @@ defaults:
       config: Vagrantfile
   devices:
     iosv:
-      interface_name: GigabitEthernet0/%d
+      interface_name: GigabitEthernet0/{ifindex}
       libvirt.image: x
 
 nodes:
